Hoist some pointer chasing out of the inner loop for a slight speed-up

+++ src/eyes/eyes.h
@@ -14,7 +14,7 @@ struct Image {
   const uint16_t width{};
   const uint16_t height{};
 
-  inline uint16_t get(uint16_t x, uint16_t y) const __attribute__((always_inline)) {
+  inline uint16_t get(uint32_t x, uint32_t y) const __attribute__((always_inline)) {
     return data[y * width + x];
   }
 };
